Added MTU support to micronets/devices and a NetReach agent config variable (#80)
In order to support tunneling inter-AP traffic over vxlans on a network with 1500-byte MTUs the MTU that NetReach devices connect with needs to be reduced so the addition of the vxlan headers doesn't exceed the 1500-byte limit.

The DHCP "Interface MTU Option" is supposed to accomplish this proactively - and avoid having to resort to Path MTU Discovery - which can cause some latency during connection establishment. 

This change adds support for an (optional) "mtu" field to both micronet and device definitions. And the dnsmasq adapter will now pick this up and set the DHCP "Interface MTU Option" (option 26) accordingly at the subnet level (for micronets) and/or the device level. 

Tested this with NetReach inter-AP tunnels. The latency that was occurring to perform inter-AP MTU path discovery is gone and iperf3 tests show no loss of frames.